fixed UI formatting
