+++ public/js/resources_js_views_Product_vue.js
@@ -85,9 +85,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _public_images_icon_plus_svg__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../../../public/images/icon-plus.svg */ "./public/images/icon-plus.svg");
 /* harmony import */ var _public_images_icon_minus_svg__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../../../public/images/icon-minus.svg */ "./public/images/icon-minus.svg");
 /* harmony import */ var _public_images_icon_cart_svg__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../public/images/icon-cart.svg */ "./public/images/icon-cart.svg");
-/* harmony import */ var vue3_toastify__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! vue3-toastify */ "./node_modules/vue3-toastify/dist/index.mjs");
-/* harmony import */ var vue3_toastify_dist_index_css__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! vue3-toastify/dist/index.css */ "./node_modules/vue3-toastify/dist/index.css");
-
+/* harmony import */ var vue3_toastify_dist_index_css__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! vue3-toastify/dist/index.css */ "./node_modules/vue3-toastify/dist/index.css");
 
 
 
@@ -124,18 +122,22 @@ __webpack_require__.r(__webpack_exports__);
   components: {},
   data: function data() {
     return {
-      nextIcon: _public_images_icon_next_svg__WEBPACK_IMPORTED_MODULE_0__["default"],
-      prevIcon: _public_images_icon_previous_svg__WEBPACK_IMPORTED_MODULE_1__["default"],
-      currentIndex: 0
+      currentIndex: 0,
+      images: [{
+        isactive: false
+      }, {
+        isactive: true
+      }, {
+        isactive: false
+      }, {
+        isactive: false
+      }]
     };
   },
   mounted: function mounted() {},
   methods: {
-    next: function next() {
-      this.currentIndex += 1;
-    },
-    prev: function prev() {
-      this.currentIndex -= 1;
+    selectImage: function selectImage(selection) {
+      this.currentIndex = selection;
     }
   },
   computed: {
@@ -143,7 +145,7 @@ __webpack_require__.r(__webpack_exports__);
       if (this.$store.state.product.images) {
         return this.$store.state.product.images[Math.abs(this.currentIndex) % this.$store.state.product.images.length];
       } else {
-        // should use a placeholder image
+        // TODO - should use a placeholder image
         return null;
       }
     }
@@ -234,17 +236,23 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export */ });
 /* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ "./node_modules/vue/dist/vue.esm-bundler.js");
 
-var _hoisted_1 = {
+var _hoisted_1 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", {
+  "class": "row d-none d-md-block",
+  style: {
+    "height": "10vw"
+  }
+}, null, -1 /* HOISTED */);
+var _hoisted_2 = {
   "class": "row"
 };
-var _hoisted_2 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", {
-  "class": "col-md-1"
+var _hoisted_3 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", {
+  "class": "d-none d-md-block col-md-1"
 }, null, -1 /* HOISTED */);
-var _hoisted_3 = {
-  "class": "p-4 col-md-5"
+var _hoisted_4 = {
+  "class": "p-4 p-md-0 ps-md-4 col-md-5"
 };
-var _hoisted_4 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", {
-  "class": "col-md-1"
+var _hoisted_5 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", {
+  "class": "d-none d-md-block col-md-1"
 }, null, -1 /* HOISTED */);
 
 function render(_ctx, _cache, $props, $setup, $data, $options) {
@@ -252,11 +260,11 @@ function render(_ctx, _cache, $props, $setup, $data, $options) {
   var _component_ImageSectionMD = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("ImageSectionMD");
   var _component_TextSection = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("TextSection");
   var _component_CartSection = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("CartSection");
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", _hoisted_1, [_hoisted_2, (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_ImageSectionXS, {
+  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", null, [_hoisted_1, (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_2, [_hoisted_3, (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_ImageSectionXS, {
     "class": "col d-md-none"
   }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_ImageSectionMD, {
-    "class": "col-5 d-none d-md-block"
-  }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_3, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_TextSection), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_CartSection)]), _hoisted_4]);
+    "class": "col-5 d-none d-md-block p-0 pe-md-4"
+  }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_4, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_TextSection), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_CartSection)]), _hoisted_5])]);
 }
 
 /***/ }),
@@ -328,7 +336,7 @@ var _hoisted_17 = ["src"];
 var _hoisted_18 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)(" Add to cart ");
 function render(_ctx, _cache, $props, $setup, $data, $options) {
   var _ctx$$store$state$pro, _ctx$$store$state$pro2, _ctx$$store$state$pro3, _ctx$$store$state$pro4, _ctx$$store$state$pro5, _ctx$$store$state$pro6, _ctx$$store$state$pro7;
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" if discount, display discounted price, discount and full price "), (_ctx$$store$state$pro = _ctx.$store.state.product.price) !== null && _ctx$$store$state$pro !== void 0 && _ctx$$store$state$pro.discounted ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", _hoisted_1, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_2, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_3, " $" + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro2 = _ctx.$store.state.product.price) === null || _ctx$$store$state$pro2 === void 0 ? void 0 : _ctx$$store$state$pro2.discounted) + ".00 ", 1 /* TEXT */), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_4, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_5, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_6, (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(((_ctx$$store$state$pro3 = _ctx.$store.state.product.discount) === null || _ctx$$store$state$pro3 === void 0 ? void 0 : _ctx$$store$state$pro3.type) == 'percent' ? '' : '$') + " " + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro4 = _ctx.$store.state.product.discount) === null || _ctx$$store$state$pro4 === void 0 ? void 0 : _ctx$$store$state$pro4.amount) + " " + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(((_ctx$$store$state$pro5 = _ctx.$store.state.product.discount) === null || _ctx$$store$state$pro5 === void 0 ? void 0 : _ctx$$store$state$pro5.type) == 'percent' ? '%' : ''), 1 /* TEXT */)])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_7, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_8, (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro6 = _ctx.$store.state.product.price) === null || _ctx$$store$state$pro6 === void 0 ? void 0 : _ctx$$store$state$pro6.full) + ".00 ", 1 /* TEXT */)])])) : ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, {
+  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" if discount, display discounted price, discount and full price "), (_ctx$$store$state$pro = _ctx.$store.state.product.price) !== null && _ctx$$store$state$pro !== void 0 && _ctx$$store$state$pro.discounted ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", _hoisted_1, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_2, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_3, " $" + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro2 = _ctx.$store.state.product.price) === null || _ctx$$store$state$pro2 === void 0 ? void 0 : _ctx$$store$state$pro2.discounted) + ".00 ", 1 /* TEXT */), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_4, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_5, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_6, (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(((_ctx$$store$state$pro3 = _ctx.$store.state.product.discount) === null || _ctx$$store$state$pro3 === void 0 ? void 0 : _ctx$$store$state$pro3.type) == 'percent' ? '' : '$') + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro4 = _ctx.$store.state.product.discount) === null || _ctx$$store$state$pro4 === void 0 ? void 0 : _ctx$$store$state$pro4.amount) + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(((_ctx$$store$state$pro5 = _ctx.$store.state.product.discount) === null || _ctx$$store$state$pro5 === void 0 ? void 0 : _ctx$$store$state$pro5.type) == 'percent' ? '%' : ''), 1 /* TEXT */)])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_7, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_8, (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro6 = _ctx.$store.state.product.price) === null || _ctx$$store$state$pro6 === void 0 ? void 0 : _ctx$$store$state$pro6.full) + ".00 ", 1 /* TEXT */)])])) : ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, {
     key: 1
   }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" if no discount, just display full price "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_9, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("p", _hoisted_10, " $" + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)((_ctx$$store$state$pro7 = _ctx.$store.state.product.price) === null || _ctx$$store$state$pro7 === void 0 ? void 0 : _ctx$$store$state$pro7.full) + ".00 ", 1 /* TEXT */)])])], 2112 /* STABLE_FRAGMENT, DEV_ROOT_FRAGMENT */)), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_11, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_12, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_13, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("a", {
     type: "button",
@@ -349,7 +357,7 @@ function render(_ctx, _cache, $props, $setup, $data, $options) {
     onClick: _cache[2] || (_cache[2] = function ($event) {
       return _ctx.$store.commit('incrementQuantity');
     })
-  }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" TODO - vertically center the plus/minus buttons "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+  }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
     src: $data.plusIcon
   }, null, 8 /* PROPS */, _hoisted_15)])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_16, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
     "class": "btn btn-add-to-cart w-100 h-100",
@@ -360,7 +368,7 @@ function render(_ctx, _cache, $props, $setup, $data, $options) {
         quantity: _ctx.$store.state.quantity
       });
     })
-  }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" TODO - vertically center the icon "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+  }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
     src: $data.cartIcon,
     "class": "cart-icon me-2"
   }, null, 8 /* PROPS */, _hoisted_17), _hoisted_18])])])], 64 /* STABLE_FRAGMENT */);
@@ -383,49 +391,35 @@ __webpack_require__.r(__webpack_exports__);
 var _withScopeId = function _withScopeId(n) {
   return (0,vue__WEBPACK_IMPORTED_MODULE_0__.pushScopeId)("data-v-12c7812a"), n = n(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.popScopeId)(), n;
 };
-var _hoisted_1 = {
-  "class": "position-relative"
+var _hoisted_1 = ["src"];
+var _hoisted_2 = {
+  "class": "row"
 };
-var _hoisted_2 = ["src"];
 var _hoisted_3 = {
-  "class": "position-absolute top-50 translate-middle-y d-flex justify-content-between w-100 px-3"
+  "class": "col d-flex justify-content-center",
+  key: "index"
 };
-var _hoisted_4 = ["src"];
+var _hoisted_4 = ["onClick"];
 var _hoisted_5 = ["src"];
 function render(_ctx, _cache, $props, $setup, $data, $options) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(vue__WEBPACK_IMPORTED_MODULE_0__.TransitionGroup, {
-    name: "fade",
-    tag: "div"
-  }, {
-    "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
-      return [((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(true), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, (0,vue__WEBPACK_IMPORTED_MODULE_0__.renderList)([$data.currentIndex], function (i) {
-        return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", {
-          key: i
-        }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_1, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
-          "class": "cropped",
-          src: $options.currentImg
-        }, null, 8 /* PROPS */, _hoisted_2), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_3, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" TODO - center icon in buttons and move them in from the edge"), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
-          type: "button",
-          "class": "btn btn-circle btn-light d-flex align-items-center justify-content-center",
-          onClick: _cache[0] || (_cache[0] = function () {
-            return $options.prev && $options.prev.apply($options, arguments);
-          })
-        }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
-          src: $data.prevIcon
-        }, null, 8 /* PROPS */, _hoisted_4)])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
-          type: "button",
-          "class": "btn btn-circle btn-light d-flex align-items-center justify-content-center",
-          onClick: _cache[1] || (_cache[1] = function () {
-            return $options.next && $options.next.apply($options, arguments);
-          })
-        }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
-          src: $data.nextIcon
-        }, null, 8 /* PROPS */, _hoisted_5)])])])])]);
-      }), 128 /* KEYED_FRAGMENT */))];
-    }),
-
-    _: 1 /* STABLE */
-  })]);
+  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+    "class": "cropped mb-4",
+    src: $options.currentImg
+  }, null, 8 /* PROPS */, _hoisted_1), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_2, [((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(true), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, (0,vue__WEBPACK_IMPORTED_MODULE_0__.renderList)(this.$store.state.product.images, function (image, index) {
+    return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", _hoisted_3, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("a", {
+      "class": (0,vue__WEBPACK_IMPORTED_MODULE_0__.normalizeClass)(["image-anchor", {
+        'selected-image-anchor': $data.currentIndex == index
+      }]),
+      onClick: function onClick($event) {
+        return $options.selectImage(index);
+      }
+    }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+      "class": (0,vue__WEBPACK_IMPORTED_MODULE_0__.normalizeClass)(["image", {
+        'selected-image': $data.currentIndex == index
+      }]),
+      src: image
+    }, null, 10 /* CLASS, PROPS */, _hoisted_5)], 10 /* CLASS, PROPS */, _hoisted_4)]);
+  }), 128 /* KEYED_FRAGMENT */))])]);
 }
 
 /***/ }),
@@ -455,39 +449,26 @@ var _hoisted_3 = {
 var _hoisted_4 = ["src"];
 var _hoisted_5 = ["src"];
 function render(_ctx, _cache, $props, $setup, $data, $options) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(vue__WEBPACK_IMPORTED_MODULE_0__.TransitionGroup, {
-    name: "fade",
-    tag: "div"
-  }, {
-    "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
-      return [((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(true), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, (0,vue__WEBPACK_IMPORTED_MODULE_0__.renderList)([$data.currentIndex], function (i) {
-        return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", {
-          key: i
-        }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_1, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
-          "class": "cropped",
-          src: $options.currentImg
-        }, null, 8 /* PROPS */, _hoisted_2), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_3, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" TODO - center icon in buttons and move them in from the edge"), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
-          type: "button",
-          "class": "btn btn-circle btn-light d-flex align-items-center justify-content-center",
-          onClick: _cache[0] || (_cache[0] = function () {
-            return $options.prev && $options.prev.apply($options, arguments);
-          })
-        }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
-          src: $data.prevIcon
-        }, null, 8 /* PROPS */, _hoisted_4)])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
-          type: "button",
-          "class": "btn btn-circle btn-light d-flex align-items-center justify-content-center",
-          onClick: _cache[1] || (_cache[1] = function () {
-            return $options.next && $options.next.apply($options, arguments);
-          })
-        }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
-          src: $data.nextIcon
-        }, null, 8 /* PROPS */, _hoisted_5)])])])])]);
-      }), 128 /* KEYED_FRAGMENT */))];
-    }),
-
-    _: 1 /* STABLE */
-  })]);
+  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_1, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+    "class": "cropped",
+    src: $options.currentImg
+  }, null, 8 /* PROPS */, _hoisted_2), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_3, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
+    type: "button",
+    "class": "btn btn-circle btn-light d-flex align-items-center justify-content-center",
+    onClick: _cache[0] || (_cache[0] = function () {
+      return $options.prev && $options.prev.apply($options, arguments);
+    })
+  }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+    src: $data.prevIcon
+  }, null, 8 /* PROPS */, _hoisted_4)])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", null, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("button", {
+    type: "button",
+    "class": "btn btn-circle btn-light d-flex align-items-center justify-content-center",
+    onClick: _cache[1] || (_cache[1] = function () {
+      return $options.next && $options.next.apply($options, arguments);
+    })
+  }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("img", {
+    src: $data.nextIcon
+  }, null, 8 /* PROPS */, _hoisted_5)])])])])]);
 }
 
 /***/ }),
@@ -524,29 +505,6 @@ function render(_ctx, _cache, $props, $setup, $data, $options) {
 
 /***/ }),
 
-/***/ "./node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!./node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./node_modules/vue3-toastify/dist/index.css":
-/*!***********************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!./node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./node_modules/vue3-toastify/dist/index.css ***!
-  \***********************************************************************************************************************************************************************************/
-/***/ ((module, __webpack_exports__, __webpack_require__) => {
-
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
-/* harmony export */ });
-/* harmony import */ var _css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../css-loader/dist/runtime/api.js */ "./node_modules/css-loader/dist/runtime/api.js");
-/* harmony import */ var _css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_0__);
-// Imports
-
-var ___CSS_LOADER_EXPORT___ = _css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_0___default()(function(i){return i[1]});
-// Module
-___CSS_LOADER_EXPORT___.push([module.id, ":root{--toastify-color-light:#fff;--toastify-color-dark:#121212;--toastify-color-info:#3498db;--toastify-color-success:#07bc0c;--toastify-color-warning:#f1c40f;--toastify-color-error:#e74c3c;--toastify-color-transparent:hsla(0,0%,100%,.7);--toastify-icon-color-info:var(--toastify-color-info);--toastify-icon-color-success:var(--toastify-color-success);--toastify-icon-color-warning:var(--toastify-color-warning);--toastify-icon-color-error:var(--toastify-color-error);--toastify-toast-width:320px;--toastify-toast-background:#fff;--toastify-toast-min-height:64px;--toastify-toast-max-height:800px;--toastify-font-family:sans-serif;--toastify-z-index:9999;--toastify-text-color-light:#757575;--toastify-text-color-dark:#fff;--toastify-text-color-info:#fff;--toastify-text-color-success:#fff;--toastify-text-color-warning:#fff;--toastify-text-color-error:#fff;--toastify-spinner-color:#616161;--toastify-spinner-color-empty-area:#e0e0e0;--toastify-color-progress-light:linear-gradient(90deg,#4cd964,#5ac8fa,#007aff,#34aadc,#5856d6,#ff2d55);--toastify-color-progress-dark:#bb86fc;--toastify-color-progress-info:var(--toastify-color-info);--toastify-color-progress-success:var(--toastify-color-success);--toastify-color-progress-warning:var(--toastify-color-warning);--toastify-color-progress-error:var(--toastify-color-error);--toastify-color-progress-colored:#ddd}.Toastify__toast-container{box-sizing:border-box;color:#fff;padding:4px;position:fixed;transform:translate3d(0,0,var(--toastify-z-index) px);width:var(--toastify-toast-width);z-index:var(--toastify-z-index)}.Toastify__toast-container--top-left{left:1em;top:1em}.Toastify__toast-container--top-center{left:50%;top:1em;transform:translateX(-50%)}.Toastify__toast-container--top-right{right:1em;top:1em}.Toastify__toast-container--bottom-left{bottom:1em;left:1em}.Toastify__toast-container--bottom-center{bottom:1em;left:50%;transform:translateX(-50%)}.Toastify__toast-container--bottom-right{bottom:1em;right:1em}@media only screen and (max-width:480px){.Toastify__toast-container{left:0;margin:0;padding:0;width:100vw}.Toastify__toast-container--top-center,.Toastify__toast-container--top-left,.Toastify__toast-container--top-right{top:0;transform:translateX(0)}.Toastify__toast-container--bottom-center,.Toastify__toast-container--bottom-left,.Toastify__toast-container--bottom-right{bottom:0;transform:translateX(0)}.Toastify__toast-container--rtl{left:auto;right:0}}.Toastify__toast{border-radius:4px;box-shadow:0 1px 10px 0 rgba(0,0,0,.1),0 2px 15px 0 rgba(0,0,0,.05);box-sizing:border-box;cursor:pointer;direction:ltr;display:flex;font-family:var(--toastify-font-family);justify-content:space-between;margin-bottom:1rem;max-height:var(--toastify-toast-max-height);min-height:var(--toastify-toast-min-height);overflow:hidden;padding:8px;position:relative;z-index:0}.Toastify__toast--rtl{direction:rtl}.Toastify__toast-body{align-items:center;display:flex;flex:1 1 auto;margin:auto 0;padding:6px;white-space:pre-wrap}.Toastify__toast-body>div:last-child{flex:1}.Toastify__toast-icon{display:flex;flex-shrink:0;margin-inline-end:10px;width:20px}.Toastify--animate{animation-duration:.7s;animation-fill-mode:both}.Toastify--animate-icon{animation-duration:.3s;animation-fill-mode:both}@media only screen and (max-width:480px){.Toastify__toast{border-radius:0;margin-bottom:0}}.Toastify__toast-theme--dark{background:var(--toastify-color-dark);color:var(--toastify-text-color-dark)}.Toastify__toast-theme--colored.Toastify__toast--default,.Toastify__toast-theme--light{background:var(--toastify-color-light);color:var(--toastify-text-color-light)}.Toastify__toast-theme--colored.Toastify__toast--info{background:var(--toastify-color-info);color:var(--toastify-text-color-info)}.Toastify__toast-theme--colored.Toastify__toast--success{background:var(--toastify-color-success);color:var(--toastify-text-color-success)}.Toastify__toast-theme--colored.Toastify__toast--warning{background:var(--toastify-color-warning);color:var(--toastify-text-color-warning)}.Toastify__toast-theme--colored.Toastify__toast--error{background:var(--toastify-color-error);color:var(--toastify-text-color-error)}.Toastify__progress-bar-theme--light{background:var(--toastify-color-progress-light)}.Toastify__progress-bar-theme--dark{background:var(--toastify-color-progress-dark)}.Toastify__progress-bar--info{background:var(--toastify-color-progress-info)}.Toastify__progress-bar--success{background:var(--toastify-color-progress-success)}.Toastify__progress-bar--warning{background:var(--toastify-color-progress-warning)}.Toastify__progress-bar--error{background:var(--toastify-color-progress-error)}.Toastify__progress-bar-theme--colored.Toastify__progress-bar--default{background:var(--toastify-color-progress-colored)}.Toastify__progress-bar-theme--colored.Toastify__progress-bar--error,.Toastify__progress-bar-theme--colored.Toastify__progress-bar--info,.Toastify__progress-bar-theme--colored.Toastify__progress-bar--success,.Toastify__progress-bar-theme--colored.Toastify__progress-bar--warning{background:var(--toastify-color-transparent)}.Toastify__close-button{align-self:flex-start;background:transparent;border:none;color:#fff;cursor:pointer;opacity:.7;outline:none;padding:0;transition:.3s ease}.Toastify__close-button--light{color:#000;opacity:.3}.Toastify__close-button>svg{fill:currentcolor;height:16px;width:14px}.Toastify__close-button:focus,.Toastify__close-button:hover{opacity:1}@keyframes Toastify__trackProgress{0%{transform:scaleX(1)}to{transform:scaleX(0)}}.Toastify__progress-bar{bottom:0;height:5px;left:0;opacity:.7;position:absolute;transform-origin:left;width:100%;z-index:var(--toastify-z-index)}.Toastify__progress-bar--animated{animation:Toastify__trackProgress linear 1 forwards}.Toastify__progress-bar--controlled{transition:transform .2s}.Toastify__progress-bar--rtl{left:auto;right:0;transform-origin:right}.Toastify__spinner{animation:Toastify__spin .65s linear infinite;border:2px solid;border-color:var(--toastify-spinner-color-empty-area);border-radius:100%;border-right-color:var(--toastify-spinner-color);box-sizing:border-box;height:20px;width:20px}@keyframes Toastify__bounceInRight{0%,60%,75%,90%,to{animation-timing-function:cubic-bezier(.215,.61,.355,1)}0%{opacity:0;transform:translate3d(3000px,0,0)}60%{opacity:1;transform:translate3d(-25px,0,0)}75%{transform:translate3d(10px,0,0)}90%{transform:translate3d(-5px,0,0)}to{transform:none}}@keyframes Toastify__bounceOutRight{20%{opacity:1;transform:translate3d(-20px,0,0)}to{opacity:0;transform:translate3d(2000px,0,0)}}@keyframes Toastify__bounceInLeft{0%,60%,75%,90%,to{animation-timing-function:cubic-bezier(.215,.61,.355,1)}0%{opacity:0;transform:translate3d(-3000px,0,0)}60%{opacity:1;transform:translate3d(25px,0,0)}75%{transform:translate3d(-10px,0,0)}90%{transform:translate3d(5px,0,0)}to{transform:none}}@keyframes Toastify__bounceOutLeft{20%{opacity:1;transform:translate3d(20px,0,0)}to{opacity:0;transform:translate3d(-2000px,0,0)}}@keyframes Toastify__bounceInUp{0%,60%,75%,90%,to{animation-timing-function:cubic-bezier(.215,.61,.355,1)}0%{opacity:0;transform:translate3d(0,3000px,0)}60%{opacity:1;transform:translate3d(0,-20px,0)}75%{transform:translate3d(0,10px,0)}90%{transform:translate3d(0,-5px,0)}to{transform:translateZ(0)}}@keyframes Toastify__bounceOutUp{20%{transform:translate3d(0,-10px,0)}40%,45%{opacity:1;transform:translate3d(0,20px,0)}to{opacity:0;transform:translate3d(0,-2000px,0)}}@keyframes Toastify__bounceInDown{0%,60%,75%,90%,to{animation-timing-function:cubic-bezier(.215,.61,.355,1)}0%{opacity:0;transform:translate3d(0,-3000px,0)}60%{opacity:1;transform:translate3d(0,25px,0)}75%{transform:translate3d(0,-10px,0)}90%{transform:translate3d(0,5px,0)}to{transform:none}}@keyframes Toastify__bounceOutDown{20%{transform:translate3d(0,10px,0)}40%,45%{opacity:1;transform:translate3d(0,-20px,0)}to{opacity:0;transform:translate3d(0,2000px,0)}}.Toastify__bounce-enter--bottom-left,.Toastify__bounce-enter--top-left{animation-name:Toastify__bounceInLeft}.Toastify__bounce-enter--bottom-right,.Toastify__bounce-enter--top-right{animation-name:Toastify__bounceInRight}.Toastify__bounce-enter--top-center{animation-name:Toastify__bounceInDown}.Toastify__bounce-enter--bottom-center{animation-name:Toastify__bounceInUp}.Toastify__bounce-exit--bottom-left,.Toastify__bounce-exit--top-left{animation-name:Toastify__bounceOutLeft}.Toastify__bounce-exit--bottom-right,.Toastify__bounce-exit--top-right{animation-name:Toastify__bounceOutRight}.Toastify__bounce-exit--top-center{animation-name:Toastify__bounceOutUp}.Toastify__bounce-exit--bottom-center{animation-name:Toastify__bounceOutDown}@keyframes Toastify__zoomIn{0%{opacity:0;transform:scale3d(.3,.3,.3)}50%{opacity:1}}@keyframes Toastify__zoomOut{0%{opacity:1}50%{opacity:0;transform:scale3d(.3,.3,.3)}to{opacity:0}}.Toastify__zoom-enter{animation-name:Toastify__zoomIn}.Toastify__zoom-exit{animation-name:Toastify__zoomOut}@keyframes Toastify__flipIn{0%{animation-timing-function:ease-in;opacity:0;transform:perspective(400px) rotateX(90deg)}40%{animation-timing-function:ease-in;transform:perspective(400px) rotateX(-20deg)}60%{opacity:1;transform:perspective(400px) rotateX(10deg)}80%{transform:perspective(400px) rotateX(-5deg)}to{transform:perspective(400px)}}@keyframes Toastify__flipOut{0%{transform:perspective(400px)}30%{opacity:1;transform:perspective(400px) rotateX(-20deg)}to{opacity:0;transform:perspective(400px) rotateX(90deg)}}.Toastify__flip-enter{animation-name:Toastify__flipIn}.Toastify__flip-exit{animation-name:Toastify__flipOut}@keyframes Toastify__slideInRight{0%{transform:translate3d(110%,0,0);visibility:visible}to{transform:translateZ(0)}}@keyframes Toastify__slideInLeft{0%{transform:translate3d(-110%,0,0);visibility:visible}to{transform:translateZ(0)}}@keyframes Toastify__slideInUp{0%{transform:translate3d(0,110%,0);visibility:visible}to{transform:translateZ(0)}}@keyframes Toastify__slideInDown{0%{transform:translate3d(0,-110%,0);visibility:visible}to{transform:translateZ(0)}}@keyframes Toastify__slideOutRight{0%{transform:translateZ(0)}to{transform:translate3d(110%,0,0);visibility:hidden}}@keyframes Toastify__slideOutLeft{0%{transform:translateZ(0)}to{transform:translate3d(-110%,0,0);visibility:hidden}}@keyframes Toastify__slideOutDown{0%{transform:translateZ(0)}to{transform:translate3d(0,500px,0);visibility:hidden}}@keyframes Toastify__slideOutUp{0%{transform:translateZ(0)}to{transform:translate3d(0,-500px,0);visibility:hidden}}.Toastify__slide-enter--bottom-left,.Toastify__slide-enter--top-left{animation-name:Toastify__slideInLeft}.Toastify__slide-enter--bottom-right,.Toastify__slide-enter--top-right{animation-name:Toastify__slideInRight}.Toastify__slide-enter--top-center{animation-name:Toastify__slideInDown}.Toastify__slide-enter--bottom-center{animation-name:Toastify__slideInUp}.Toastify__slide-exit--bottom-left,.Toastify__slide-exit--top-left{animation-name:Toastify__slideOutLeft}.Toastify__slide-exit--bottom-right,.Toastify__slide-exit--top-right{animation-name:Toastify__slideOutRight}.Toastify__slide-exit--top-center{animation-name:Toastify__slideOutUp}.Toastify__slide-exit--bottom-center{animation-name:Toastify__slideOutDown}@keyframes Toastify__spin{0%{transform:rotate(0deg)}to{transform:rotate(1turn)}}", ""]);
-// Exports
-/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);
-
-
-/***/ }),
-
 /***/ "./node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!./node_modules/vue-loader/dist/stylePostLoader.js!./node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/views/components/CartSection.vue?vue&type=style&index=0&id=6e2037cb&scoped=true&lang=css":
 /*!************************************************************************************************************************************************************************************************************************************************************************************************************************************************************!*\
   !*** ./node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!./node_modules/vue-loader/dist/stylePostLoader.js!./node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/views/components/CartSection.vue?vue&type=style&index=0&id=6e2037cb&scoped=true&lang=css ***!
@@ -563,7 +521,7 @@ __webpack_require__.r(__webpack_exports__);
 
 var ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_0___default()(function(i){return i[1]});
 // Module
-___CSS_LOADER_EXPORT___.push([module.id, "\n.product-price[data-v-6e2037cb]{\n    font-size: 28px;\n    font-weight: bold;\n}\n.bg-pale-orange[data-v-6e2037cb] {\n    background-color: hsl(25, 100%, 94%);\n    width: -moz-fit-content;\n    width: fit-content;\n}\n.full-price-text[data-v-6e2037cb] {\n    font-size: 14px;\n    font-weight: 600;\n    color: hsl(220, 14%, 75%);\n}\ninput[type=\"number\"][data-v-6e2037cb] {\n  -webkit-appearance: textfield;\n  -moz-appearance: textfield;\n  appearance: textfield;\n  background-color: hsl(223, 64%, 98%);\n  border: 0;\n  width: 100px;\n  font-weight: 600;\n  text-align: center;\n}\ninput[type=number][data-v-6e2037cb]::-webkit-inner-spin-button,\ninput[type=number][data-v-6e2037cb]::-webkit-outer-spin-button {\n  -webkit-appearance: none;\n}\n.quantity-area[data-v-6e2037cb] {\n    background-color: hsl(223, 64%, 98%);\n    border-radius: 10px;\n    height: 50px;\n}\n.product-quantity-input[data-v-6e2037cb]:focus {\n    outline: 0;\n}\n.btn-add-to-cart[data-v-6e2037cb] {\n    background-color: hsl(26, 100%, 55%);\n    color: white;\n    font-weight: 600;\n    border-radius: 10px;\n}\n.cart-icon[data-v-6e2037cb] {\n    filter: brightness(0) invert(1);\n    width: 15px;\n    height: 15px;\n}\n\n", ""]);
+___CSS_LOADER_EXPORT___.push([module.id, "\n.product-price[data-v-6e2037cb]{\n    font-size: 28px;\n    font-weight: bold;\n}\n.bg-pale-orange[data-v-6e2037cb] {\n    background-color: hsl(25, 100%, 94%);\n    width: -moz-fit-content;\n    width: fit-content;\n}\n.orange-bold-text[data-v-6e2037cb] {\n    color: hsl(26, 100%, 55%);\n    font-size: 14px;\n    font-weight: 600\n}\n.full-price-text[data-v-6e2037cb] {\n    font-size: 14px;\n    font-weight: 600;\n    color: hsl(220, 14%, 75%);\n}\ninput[type=\"number\"][data-v-6e2037cb] {\n  -webkit-appearance: textfield;\n  -moz-appearance: textfield;\n  appearance: textfield;\n  background-color: hsl(223, 64%, 98%);\n  border: 0;\n  width: 100px;\n  font-weight: 600;\n  text-align: center;\n}\ninput[type=number][data-v-6e2037cb]::-webkit-inner-spin-button,\ninput[type=number][data-v-6e2037cb]::-webkit-outer-spin-button {\n  -webkit-appearance: none;\n}\n.quantity-area[data-v-6e2037cb] {\n    background-color: hsl(223, 64%, 98%);\n    border-radius: 10px;\n    height: 50px;\n}\n.product-quantity-input[data-v-6e2037cb]:focus {\n    outline: 0;\n}\n.btn-add-to-cart[data-v-6e2037cb] {\n    background-color: hsl(26, 100%, 55%);\n    color: white;\n    font-weight: 600;\n    border-radius: 10px;\n}\n.cart-icon[data-v-6e2037cb] {\n    filter: brightness(0) invert(1);\n    width: 15px;\n    height: 15px;\n}\n\n", ""]);
 // Exports
 /* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);
 
@@ -586,7 +544,7 @@ __webpack_require__.r(__webpack_exports__);
 
 var ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_0___default()(function(i){return i[1]});
 // Module
-___CSS_LOADER_EXPORT___.push([module.id, "\n.img-responsive[data-v-12c7812a] {\n    width: 100%;\n    height: auto;\n}\n.btn-circle[data-v-12c7812a] { \n    width: 40px; \n    height: 40px; \n    border-radius: 20px;\n}\n.cropped[data-v-12c7812a] {\n    -o-object-fit: cover;\n       object-fit: cover;\n    height: 80vw;\n    width: 100%;\n}\n\n", ""]);
+___CSS_LOADER_EXPORT___.push([module.id, "\n.img-responsive[data-v-12c7812a] {\n    width: 100%;\n    height: auto;\n}\n.btn-circle[data-v-12c7812a] { \n    width: 40px; \n    height: 40px; \n    border-radius: 20px;\n}\n.cropped[data-v-12c7812a] {\n    -o-object-fit: cover;\n       object-fit: cover;\n    height: auto;\n    width: 100%;\n    border-radius: 10px;\n}\n.image[data-v-12c7812a] {\n    -o-object-fit: cover;\n       object-fit: cover;\n    height: auto;\n    width: 60px;\n    border-radius: 5px;\n}\n.selected-image[data-v-12c7812a] {\n    opacity: 0.2;\n}\n.image-anchor[data-v-12c7812a] {\n    cursor: pointer;\n    border: 0px;\n    border-radius: 7px;\n}\n.selected-image-anchor[data-v-12c7812a] {\n    border: 2px solid hsl(26, 100%, 55%);\n}\n\n", ""]);
 // Exports
 /* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);
 
@@ -706,35 +664,6 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
 /* harmony export */ });
 /* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = ("/images/icon-previous.svg?80616402c611cc0a6651d0b751da8a9b");
-
-/***/ }),
-
-/***/ "./node_modules/vue3-toastify/dist/index.css":
-/*!***************************************************!*\
-  !*** ./node_modules/vue3-toastify/dist/index.css ***!
-  \***************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
-/* harmony export */ });
-/* harmony import */ var _style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! !../../style-loader/dist/runtime/injectStylesIntoStyleTag.js */ "./node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js");
-/* harmony import */ var _style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0__);
-/* harmony import */ var _css_loader_dist_cjs_js_clonedRuleSet_9_use_1_postcss_loader_dist_cjs_js_clonedRuleSet_9_use_2_index_css__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! !!../../css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!../../postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./index.css */ "./node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!./node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./node_modules/vue3-toastify/dist/index.css");
-
-            
-
-var options = {};
-
-options.insert = "head";
-options.singleton = false;
-
-var update = _style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0___default()(_css_loader_dist_cjs_js_clonedRuleSet_9_use_1_postcss_loader_dist_cjs_js_clonedRuleSet_9_use_2_index_css__WEBPACK_IMPORTED_MODULE_1__["default"], options);
-
-
-
-/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (_css_loader_dist_cjs_js_clonedRuleSet_9_use_1_postcss_loader_dist_cjs_js_clonedRuleSet_9_use_2_index_css__WEBPACK_IMPORTED_MODULE_1__["default"].locals || {});
 
 /***/ }),
 
@@ -1216,1075 +1145,6 @@ __webpack_require__.r(__webpack_exports__);
 
 __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _node_modules_style_loader_dist_cjs_js_node_modules_css_loader_dist_cjs_js_clonedRuleSet_9_use_1_node_modules_vue_loader_dist_stylePostLoader_js_node_modules_postcss_loader_dist_cjs_js_clonedRuleSet_9_use_2_node_modules_vue_loader_dist_index_js_ruleSet_0_use_0_textSection_vue_vue_type_style_index_0_id_80ae4884_scoped_true_lang_css__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/style-loader/dist/cjs.js!../../../../node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!../../../../node_modules/vue-loader/dist/stylePostLoader.js!../../../../node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!../../../../node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./textSection.vue?vue&type=style&index=0&id=80ae4884&scoped=true&lang=css */ "./node_modules/style-loader/dist/cjs.js!./node_modules/css-loader/dist/cjs.js??clonedRuleSet-9.use[1]!./node_modules/vue-loader/dist/stylePostLoader.js!./node_modules/postcss-loader/dist/cjs.js??clonedRuleSet-9.use[2]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/views/components/textSection.vue?vue&type=style&index=0&id=80ae4884&scoped=true&lang=css");
-
-
-/***/ }),
-
-/***/ "./node_modules/vue3-toastify/dist/index.mjs":
-/*!***************************************************!*\
-  !*** ./node_modules/vue3-toastify/dist/index.mjs ***!
-  \***************************************************/
-/***/ ((__unused_webpack___webpack_module__, __webpack_exports__, __webpack_require__) => {
-
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   AnimationStep: () => (/* binding */ V),
-/* harmony export */   Bounce: () => (/* binding */ _e),
-/* harmony export */   Flip: () => (/* binding */ be),
-/* harmony export */   Slide: () => (/* binding */ he),
-/* harmony export */   ToastActions: () => (/* binding */ A),
-/* harmony export */   ToastifyContainer: () => (/* binding */ rt),
-/* harmony export */   Zoom: () => (/* binding */ Oe),
-/* harmony export */   addExitAnimateToNode: () => (/* binding */ ge),
-/* harmony export */   appendFromQueue: () => (/* binding */ De),
-/* harmony export */   cacheRenderInstance: () => (/* binding */ Me),
-/* harmony export */   clearContainers: () => (/* binding */ we),
-/* harmony export */   containerInstances: () => (/* binding */ x),
-/* harmony export */   "default": () => (/* binding */ dt),
-/* harmony export */   doAppend: () => (/* binding */ j),
-/* harmony export */   getAllToast: () => (/* binding */ k),
-/* harmony export */   getToast: () => (/* binding */ ke),
-/* harmony export */   globalCache: () => (/* binding */ Q),
-/* harmony export */   globalOptions: () => (/* binding */ Ce),
-/* harmony export */   queue: () => (/* binding */ N),
-/* harmony export */   removeContainer: () => (/* binding */ te),
-/* harmony export */   toast: () => (/* binding */ l),
-/* harmony export */   toastContainers: () => (/* binding */ c),
-/* harmony export */   updateGlobalOptions: () => (/* binding */ ct),
-/* harmony export */   useCssTransition: () => (/* binding */ ot)
-/* harmony export */ });
-/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ "./node_modules/vue/dist/vue.esm-bundler.js");
-
-const D = {
-  TOP_LEFT: "top-left",
-  TOP_RIGHT: "top-right",
-  TOP_CENTER: "top-center",
-  BOTTOM_LEFT: "bottom-left",
-  BOTTOM_RIGHT: "bottom-right",
-  BOTTOM_CENTER: "bottom-center"
-}, M = {
-  LIGHT: "light",
-  DARK: "dark",
-  COLORED: "colored",
-  AUTO: "auto"
-}, g = {
-  INFO: "info",
-  SUCCESS: "success",
-  WARNING: "warning",
-  ERROR: "error",
-  DEFAULT: "default"
-}, Ne = {
-  BOUNCE: "bounce",
-  SLIDE: "slide",
-  FLIP: "flip",
-  ZOOM: "zoom"
-}, ce = {
-  dangerouslyHTMLString: !1,
-  multiple: !0,
-  position: D.TOP_RIGHT,
-  autoClose: 5e3,
-  transition: "bounce",
-  hideProgressBar: !1,
-  pauseOnHover: !0,
-  pauseOnFocusLoss: !0,
-  closeOnClick: !0,
-  className: "",
-  bodyClassName: "",
-  style: {},
-  progressClassName: "",
-  progressStyle: {},
-  role: "alert",
-  theme: "light"
-}, Ie = {
-  rtl: !1,
-  newestOnTop: !1,
-  toastClassName: ""
-}, fe = {
-  ...ce,
-  ...Ie
-};
-({
-  ...ce,
-  type: g.DEFAULT
-});
-var r = /* @__PURE__ */ ((e) => (e[e.COLLAPSE_DURATION = 300] = "COLLAPSE_DURATION", e[e.DEBOUNCE_DURATION = 50] = "DEBOUNCE_DURATION", e.CSS_NAMESPACE = "Toastify", e))(r || {}), J = /* @__PURE__ */ ((e) => (e.ENTRANCE_ANIMATION_END = "d", e))(J || {});
-const _e = {
-  enter: "Toastify--animate Toastify__bounce-enter",
-  exit: "Toastify--animate Toastify__bounce-exit",
-  appendPosition: !0
-}, he = {
-  enter: "Toastify--animate Toastify__slide-enter",
-  exit: "Toastify--animate Toastify__slide-exit",
-  appendPosition: !0
-}, Oe = {
-  enter: "Toastify--animate Toastify__zoom-enter",
-  exit: "Toastify--animate Toastify__zoom-exit"
-}, be = {
-  enter: "Toastify--animate Toastify__flip-enter",
-  exit: "Toastify--animate Toastify__flip-exit"
-};
-function me(e) {
-  let t = _e;
-  if (!e || typeof e == "string")
-    switch (e) {
-      case "flip":
-        t = be;
-        break;
-      case "zoom":
-        t = Oe;
-        break;
-      case "slide":
-        t = he;
-        break;
-    }
-  else
-    t = e;
-  return t;
-}
-function Pe(e) {
-  return e.containerId || String(e.position);
-}
-const K = "will-unmount";
-function Le(e = D.TOP_RIGHT) {
-  return !!document.querySelector(".".concat(r.CSS_NAMESPACE, "__toast-container--").concat(e));
-}
-function $e(e = D.TOP_RIGHT) {
-  return "".concat(r.CSS_NAMESPACE, "__toast-container--").concat(e);
-}
-function qe(e, t, n = !1) {
-  const a = [
-    "".concat(r.CSS_NAMESPACE, "__toast-container"),
-    "".concat(r.CSS_NAMESPACE, "__toast-container--").concat(e),
-    n ? "".concat(r.CSS_NAMESPACE, "__toast-container--rtl") : null
-  ].filter(Boolean).join(" ");
-  return B(t) ? t({
-    position: e,
-    rtl: n,
-    defaultClassName: a
-  }) : "".concat(a, " ").concat(t || "");
-}
-function Be(e) {
-  var C;
-  const { position: t, containerClassName: n, rtl: a = !1, style: o = {} } = e, s = r.CSS_NAMESPACE, u = $e(t), E = document.querySelector(".".concat(s)), d = document.querySelector(".".concat(u)), p = !!d && !((C = d.className) != null && C.includes(K)), m = E || document.createElement("div"), S = document.createElement("div");
-  S.className = qe(
-    t,
-    n,
-    a
-  ), S.dataset.testid = "".concat(r.CSS_NAMESPACE, "__toast-container--").concat(t), S.id = Pe(e);
-  for (const v in o)
-    if (Object.prototype.hasOwnProperty.call(o, v)) {
-      const I = o[v];
-      S.style[v] = I;
-    }
-  return E || (m.className = r.CSS_NAMESPACE, document.body.appendChild(m)), p || m.appendChild(S), S;
-}
-function ee(e) {
-  var a, o, s;
-  const t = typeof e == "string" ? e : ((a = e.currentTarget) == null ? void 0 : a.id) || ((o = e.target) == null ? void 0 : o.id), n = document.getElementById(t);
-  n && n.removeEventListener("animationend", ee, !1);
-  try {
-    x[t].unmount(), (s = document.getElementById(t)) == null || s.remove(), delete x[t], delete c[t];
-  } catch {
-  }
-}
-const x = (0,vue__WEBPACK_IMPORTED_MODULE_0__.reactive)({});
-function Me(e, t) {
-  const n = document.getElementById(String(t));
-  n && (x[n.id] = e);
-}
-function te(e, t = !0) {
-  const n = String(e);
-  if (!x[n])
-    return;
-  const a = document.getElementById(n);
-  a && a.classList.add(K), t ? (Re(e), a && a.addEventListener("animationend", ee, !1)) : ee(n), N.items = N.items.filter((o) => o.containerId !== e);
-}
-function we(e) {
-  for (const t in x)
-    te(t, e);
-  N.items = [];
-}
-function ge(e, t) {
-  const n = document.getElementById(e.toastId);
-  if (n) {
-    let a = e;
-    a = {
-      ...a,
-      ...me(a.transition)
-    };
-    const o = a.appendPosition ? "".concat(a.exit, "--").concat(a.position) : a.exit;
-    n.className += " ".concat(o), t && t(n);
-  }
-}
-function Re(e) {
-  for (const t in c)
-    if (t === e)
-      for (const n of c[t] || [])
-        ge(n);
-}
-function Fe(e) {
-  const n = k().find((a) => a.toastId === e);
-  return n == null ? void 0 : n.containerId;
-}
-function se(e) {
-  return document.getElementById(e);
-}
-function Ue(e) {
-  const t = se(e.containerId);
-  return t && t.classList.contains(K);
-}
-function re(e) {
-  var n;
-  const t = (0,vue__WEBPACK_IMPORTED_MODULE_0__.isVNode)(e.content) ? (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(e.content.props) : null;
-  return t != null ? t : (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)((n = e.data) != null ? n : {});
-}
-function xe(e) {
-  return e ? N.items.filter((n) => n.containerId === e).length > 0 : N.items.length > 0;
-}
-function De() {
-  if (N.items.length > 0) {
-    const e = N.items.shift();
-    j(e == null ? void 0 : e.toastContent, e == null ? void 0 : e.toastProps);
-  }
-}
-const c = (0,vue__WEBPACK_IMPORTED_MODULE_0__.reactive)({}), N = (0,vue__WEBPACK_IMPORTED_MODULE_0__.reactive)({
-  items: []
-});
-function k() {
-  const e = (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(c);
-  return Object.values(e).reduce((t, n) => [...t, ...n], []);
-}
-function ke(e) {
-  return k().find((n) => n.toastId === e);
-}
-function j(e, t = {}) {
-  if (Ue(t)) {
-    const n = se(t.containerId);
-    n && n.addEventListener("animationend", ne.bind(null, e, t), !1);
-  } else
-    ne(e, t);
-}
-function ne(e, t = {}) {
-  const n = se(t.containerId);
-  n && n.removeEventListener("animationend", ne.bind(null, e, t), !1);
-  const a = c[t.containerId] || [], o = a.length > 0;
-  if (!o && !Le(t.position)) {
-    const s = Be(t), u = (0,vue__WEBPACK_IMPORTED_MODULE_0__.createApp)(rt, t);
-    u.mount(s), Me(u, s.id);
-  }
-  o && (t.position = a[0].position), (0,vue__WEBPACK_IMPORTED_MODULE_0__.nextTick)(() => {
-    t.updateId ? A.update(t) : A.add(e, t);
-  });
-}
-const A = {
-  /**
-   * add a toast
-   * @param _ ..
-   * @param opts toast props
-   */
-  add(e, t) {
-    const { containerId: n = "" } = t;
-    n && (c[n] = c[n] || [], c[n].find((a) => a.toastId === t.toastId) || setTimeout(() => {
-      var a, o;
-      t.newestOnTop ? (a = c[n]) == null || a.unshift(t) : (o = c[n]) == null || o.push(t), t.onOpen && t.onOpen(re(t));
-    }, t.delay || 0));
-  },
-  /**
-   * remove a toast
-   * @param id toastId
-   */
-  remove(e) {
-    if (e) {
-      const t = Fe(e);
-      if (t) {
-        const n = c[t];
-        let a = n.find((o) => o.toastId === e);
-        c[t] = n.filter((o) => o.toastId !== e), !c[t].length && !xe(t) && te(t, !1), De(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.nextTick)(() => {
-          a != null && a.onClose && (a.onClose(re(a)), a = void 0);
-        });
-      }
-    }
-  },
-  /**
-   * update the toast
-   * @param opts toast props
-   */
-  update(e = {}) {
-    const { containerId: t = "" } = e;
-    if (t && e.updateId) {
-      c[t] = c[t] || [];
-      const n = c[t].find((a) => a.toastId === e.toastId);
-      n && setTimeout(() => {
-        for (const a in e)
-          if (Object.prototype.hasOwnProperty.call(e, a)) {
-            const o = e[a];
-            n[a] = o;
-          }
-      }, e.delay || 0);
-    }
-  },
-  /**
-   * clear all toasts in container.
-   * @param containerId container id
-   */
-  clear(e, t = !0) {
-    e ? te(e, t) : we(t);
-  },
-  dismissCallback(e) {
-    var a;
-    const t = (a = e.currentTarget) == null ? void 0 : a.id, n = document.getElementById(t);
-    n && (n.removeEventListener("animationend", A.dismissCallback, !1), setTimeout(() => {
-      A.remove(t);
-    }));
-  },
-  dismiss(e) {
-    if (e) {
-      const t = k();
-      for (const n of t)
-        if (n.toastId === e) {
-          ge(n, (a) => {
-            a.addEventListener("animationend", A.dismissCallback, !1);
-          });
-          break;
-        }
-    }
-  }
-}, Ce = (0,vue__WEBPACK_IMPORTED_MODULE_0__.reactive)({}), Q = (0,vue__WEBPACK_IMPORTED_MODULE_0__.reactive)({});
-function ye() {
-  return Math.random().toString(36).substring(2, 9);
-}
-function He(e) {
-  return typeof e == "number" && !isNaN(e);
-}
-function ae(e) {
-  return typeof e == "string";
-}
-function B(e) {
-  return typeof e == "function";
-}
-function Y(...e) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)(...e);
-}
-function G(e) {
-  return typeof e == "object" && (!!(e != null && e.render) || !!(e != null && e.setup) || typeof (e == null ? void 0 : e.type) == "object");
-}
-function ze(e = {}) {
-  Ce["".concat(r.CSS_NAMESPACE, "-default-options")] = e;
-}
-function je() {
-  return Ce["".concat(r.CSS_NAMESPACE, "-default-options")] || fe;
-}
-function Ge() {
-  return document.documentElement.classList.contains("dark") ? "dark" : "light";
-}
-var V = /* @__PURE__ */ ((e) => (e[e.Enter = 0] = "Enter", e[e.Exit = 1] = "Exit", e))(V || {});
-const Ee = {
-  containerId: {
-    type: [String, Number],
-    required: !1,
-    default: ""
-  },
-  clearOnUrlChange: {
-    type: Boolean,
-    required: !1,
-    default: !0
-  },
-  dangerouslyHTMLString: {
-    type: Boolean,
-    required: !1,
-    default: !1
-  },
-  multiple: {
-    type: Boolean,
-    required: !1,
-    default: !0
-  },
-  limit: {
-    type: Number,
-    required: !1,
-    default: void 0
-  },
-  position: {
-    type: String,
-    required: !1,
-    default: D.TOP_LEFT
-  },
-  bodyClassName: {
-    type: String,
-    required: !1,
-    default: ""
-  },
-  autoClose: {
-    type: [Number, Boolean],
-    required: !1,
-    default: !1
-  },
-  closeButton: {
-    type: [Boolean, Function, Object],
-    required: !1,
-    default: void 0
-  },
-  transition: {
-    type: [String, Object],
-    required: !1,
-    default: "bounce"
-  },
-  hideProgressBar: {
-    type: Boolean,
-    required: !1,
-    default: !1
-  },
-  pauseOnHover: {
-    type: Boolean,
-    required: !1,
-    default: !0
-  },
-  pauseOnFocusLoss: {
-    type: Boolean,
-    required: !1,
-    default: !0
-  },
-  closeOnClick: {
-    type: Boolean,
-    required: !1,
-    default: !0
-  },
-  progress: {
-    type: Number,
-    required: !1,
-    default: void 0
-  },
-  progressClassName: {
-    type: String,
-    required: !1,
-    default: ""
-  },
-  toastStyle: {
-    type: Object,
-    required: !1,
-    default() {
-      return {};
-    }
-  },
-  progressStyle: {
-    type: Object,
-    required: !1,
-    default() {
-      return {};
-    }
-  },
-  role: {
-    type: String,
-    required: !1,
-    default: "alert"
-  },
-  theme: {
-    type: String,
-    required: !1,
-    default: M.AUTO
-  },
-  content: {
-    type: [String, Object, Function],
-    required: !1,
-    default: ""
-  },
-  toastId: {
-    type: [String, Number],
-    required: !1,
-    default: ""
-  },
-  data: {
-    type: [Object, String],
-    required: !1,
-    default() {
-      return {};
-    }
-  },
-  type: {
-    type: String,
-    required: !1,
-    default: g.DEFAULT
-  },
-  icon: {
-    type: [Boolean, String, Number, Object, Function],
-    required: !1,
-    default: void 0
-  },
-  delay: {
-    type: Number,
-    required: !1,
-    default: void 0
-  },
-  onOpen: {
-    type: Function,
-    required: !1,
-    default: void 0
-  },
-  onClose: {
-    type: Function,
-    required: !1,
-    default: void 0
-  },
-  onClick: {
-    type: Function,
-    required: !1,
-    default: void 0
-  },
-  isLoading: {
-    type: Boolean,
-    required: !1,
-    default: void 0
-  },
-  rtl: {
-    type: Boolean,
-    required: !1,
-    default: !1
-  },
-  toastClassName: {
-    type: String,
-    required: !1,
-    default: ""
-  },
-  updateId: {
-    type: [String, Number],
-    required: !1,
-    default: ""
-  }
-}, Ve = {
-  autoClose: {
-    type: [Number, Boolean],
-    required: !0
-  },
-  isRunning: {
-    type: Boolean,
-    required: !1,
-    default: void 0
-  },
-  type: {
-    type: String,
-    required: !1,
-    default: g.DEFAULT
-  },
-  theme: {
-    type: String,
-    required: !1,
-    default: M.AUTO
-  },
-  hide: {
-    type: Boolean,
-    required: !1,
-    default: void 0
-  },
-  className: {
-    type: [String, Function],
-    required: !1,
-    default: ""
-  },
-  controlledProgress: {
-    type: Boolean,
-    required: !1,
-    default: void 0
-  },
-  rtl: {
-    type: Boolean,
-    required: !1,
-    default: void 0
-  },
-  isIn: {
-    type: Boolean,
-    required: !1,
-    default: void 0
-  },
-  progress: {
-    type: Number,
-    required: !1,
-    default: void 0
-  },
-  closeToast: {
-    type: Function,
-    required: !1,
-    default: void 0
-  }
-}, Qe = /* @__PURE__ */ (0,vue__WEBPACK_IMPORTED_MODULE_0__.defineComponent)({
-  name: "ProgressBar",
-  props: Ve,
-  // @ts-ignore
-  setup(e, {
-    attrs: t
-  }) {
-    const n = (0,vue__WEBPACK_IMPORTED_MODULE_0__.ref)(), a = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => e.hide ? "true" : "false"), o = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => ({
-      ...t.style || {},
-      animationDuration: "".concat(e.autoClose === !0 ? 5e3 : e.autoClose, "ms"),
-      animationPlayState: e.isRunning ? "running" : "paused",
-      opacity: e.hide || e.autoClose === !1 ? 0 : 1,
-      transform: e.controlledProgress ? "scaleX(".concat(e.progress, ")") : "none"
-    })), s = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => ["".concat(r.CSS_NAMESPACE, "__progress-bar"), e.controlledProgress ? "".concat(r.CSS_NAMESPACE, "__progress-bar--controlled") : "".concat(r.CSS_NAMESPACE, "__progress-bar--animated"), "".concat(r.CSS_NAMESPACE, "__progress-bar-theme--").concat(e.theme), "".concat(r.CSS_NAMESPACE, "__progress-bar--").concat(e.type), e.rtl ? "".concat(r.CSS_NAMESPACE, "__progress-bar--rtl") : null].filter(Boolean).join(" ")), u = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => "".concat(s.value, " ").concat((t == null ? void 0 : t.class) || "")), E = () => {
-      n.value && (n.value.onanimationend = null, n.value.ontransitionend = null);
-    }, d = () => {
-      e.isIn && e.closeToast && e.autoClose !== !1 && (e.closeToast(), E());
-    }, p = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => e.controlledProgress ? null : d), m = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => e.controlledProgress ? d : null);
-    return (0,vue__WEBPACK_IMPORTED_MODULE_0__.watchEffect)(() => {
-      n.value && (E(), n.value.onanimationend = p.value, n.value.ontransitionend = m.value);
-    }), () => (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("div", {
-      ref: n,
-      role: "progressbar",
-      "aria-hidden": a.value,
-      "aria-label": "notification timer",
-      class: u.value,
-      style: o.value
-    }, null);
-  }
-}), We = /* @__PURE__ */ (0,vue__WEBPACK_IMPORTED_MODULE_0__.defineComponent)({
-  name: "CloseButton",
-  inheritAttrs: !1,
-  props: {
-    theme: {
-      type: String,
-      required: !1,
-      default: M.AUTO
-    },
-    type: {
-      type: String,
-      required: !1,
-      default: M.LIGHT
-    },
-    ariaLabel: {
-      type: String,
-      required: !1,
-      default: "close"
-    },
-    closeToast: {
-      type: Function,
-      required: !1,
-      default: void 0
-    }
-  },
-  setup(e) {
-    return () => (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("button", {
-      class: "".concat(r.CSS_NAMESPACE, "__close-button ").concat(r.CSS_NAMESPACE, "__close-button--").concat(e.theme),
-      type: "button",
-      onClick: (t) => {
-        t.stopPropagation(), e.closeToast && e.closeToast(t);
-      },
-      "aria-label": e.ariaLabel
-    }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("svg", {
-      "aria-hidden": "true",
-      viewBox: "0 0 14 16"
-    }, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("path", {
-      "fill-rule": "evenodd",
-      d: "M7.71 8.23l3.75 3.75-1.48 1.48-3.75-3.75-3.75 3.75L1 11.98l3.75-3.75L1 4.48 2.48 3l3.75 3.75L9.98 3l1.48 1.48-3.75 3.75z"
-    }, null)])]);
-  }
-}), Z = ({
-  theme: e,
-  type: t,
-  path: n,
-  ...a
-}) => (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("svg", (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)({
-  viewBox: "0 0 24 24",
-  width: "100%",
-  height: "100%",
-  fill: e === "colored" ? "currentColor" : "var(--toastify-icon-color-".concat(t, ")")
-}, a), [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("path", {
-  d: n
-}, null)]);
-function Ke(e) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(Z, (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)(e, {
-    path: "M23.32 17.191L15.438 2.184C14.728.833 13.416 0 11.996 0c-1.42 0-2.733.833-3.443 2.184L.533 17.448a4.744 4.744 0 000 4.368C1.243 23.167 2.555 24 3.975 24h16.05C22.22 24 24 22.044 24 19.632c0-.904-.251-1.746-.68-2.44zm-9.622 1.46c0 1.033-.724 1.823-1.698 1.823s-1.698-.79-1.698-1.822v-.043c0-1.028.724-1.822 1.698-1.822s1.698.79 1.698 1.822v.043zm.039-12.285l-.84 8.06c-.057.581-.408.943-.897.943-.49 0-.84-.367-.896-.942l-.84-8.065c-.057-.624.25-1.095.779-1.095h1.91c.528.005.84.476.784 1.1z"
-  }), null);
-}
-function Ye(e) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(Z, (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)(e, {
-    path: "M12 0a12 12 0 1012 12A12.013 12.013 0 0012 0zm.25 5a1.5 1.5 0 11-1.5 1.5 1.5 1.5 0 011.5-1.5zm2.25 13.5h-4a1 1 0 010-2h.75a.25.25 0 00.25-.25v-4.5a.25.25 0 00-.25-.25h-.75a1 1 0 010-2h1a2 2 0 012 2v4.75a.25.25 0 00.25.25h.75a1 1 0 110 2z"
-  }), null);
-}
-function Ze(e) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(Z, (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)(e, {
-    path: "M12 0a12 12 0 1012 12A12.014 12.014 0 0012 0zm6.927 8.2l-6.845 9.289a1.011 1.011 0 01-1.43.188l-4.888-3.908a1 1 0 111.25-1.562l4.076 3.261 6.227-8.451a1 1 0 111.61 1.183z"
-  }), null);
-}
-function Xe(e) {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(Z, (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)(e, {
-    path: "M11.983 0a12.206 12.206 0 00-8.51 3.653A11.8 11.8 0 000 12.207 11.779 11.779 0 0011.8 24h.214A12.111 12.111 0 0024 11.791 11.766 11.766 0 0011.983 0zM10.5 16.542a1.476 1.476 0 011.449-1.53h.027a1.527 1.527 0 011.523 1.47 1.475 1.475 0 01-1.449 1.53h-.027a1.529 1.529 0 01-1.523-1.47zM11 12.5v-6a1 1 0 012 0v6a1 1 0 11-2 0z"
-  }), null);
-}
-function Je() {
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("div", {
-    class: "".concat(r.CSS_NAMESPACE, "__spinner")
-  }, null);
-}
-const oe = {
-  info: Ye,
-  warning: Ke,
-  success: Ze,
-  error: Xe,
-  spinner: Je
-}, et = (e) => e in oe;
-function tt({
-  theme: e,
-  type: t,
-  isLoading: n,
-  icon: a
-}) {
-  let o;
-  const s = {
-    theme: e,
-    type: t
-  };
-  return n ? o = oe.spinner() : a === !1 ? o = void 0 : G(a) ? o = (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(a) : B(a) ? o = a(s) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.isVNode)(a) ? o = (0,vue__WEBPACK_IMPORTED_MODULE_0__.cloneVNode)(a, s) : ae(a) || He(a) ? o = a : et(t) && (o = oe[t](s)), o;
-}
-const nt = () => {
-};
-function at(e, t, n = r.COLLAPSE_DURATION) {
-  const { scrollHeight: a, style: o } = e, s = n;
-  requestAnimationFrame(() => {
-    o.minHeight = "initial", o.height = a + "px", o.transition = "all ".concat(s, "ms"), requestAnimationFrame(() => {
-      o.height = "0", o.padding = "0", o.margin = "0", setTimeout(t, s);
-    });
-  });
-}
-function ot(e) {
-  const t = (0,vue__WEBPACK_IMPORTED_MODULE_0__.ref)(!1), n = (0,vue__WEBPACK_IMPORTED_MODULE_0__.ref)(!1), a = (0,vue__WEBPACK_IMPORTED_MODULE_0__.ref)(!1), o = (0,vue__WEBPACK_IMPORTED_MODULE_0__.ref)(V.Enter), s = (0,vue__WEBPACK_IMPORTED_MODULE_0__.reactive)({
-    ...e,
-    appendPosition: e.appendPosition || !1,
-    collapse: typeof e.collapse > "u" ? !0 : e.collapse,
-    collapseDuration: e.collapseDuration || r.COLLAPSE_DURATION
-  }), u = s.done || nt, E = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => s.appendPosition ? "".concat(s.enter, "--").concat(s.position) : s.enter), d = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => s.appendPosition ? "".concat(s.exit, "--").concat(s.position) : s.exit), p = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => e.pauseOnHover ? {
-    onMouseenter: h,
-    onMouseleave: _
-  } : {});
-  function m() {
-    const y = E.value.split(" ");
-    C().addEventListener(
-      J.ENTRANCE_ANIMATION_END,
-      _,
-      { once: !0 }
-    );
-    const O = ($) => {
-      const F = C();
-      $.target === F && (F.dispatchEvent(new Event(J.ENTRANCE_ANIMATION_END)), F.removeEventListener("animationend", O), F.removeEventListener("animationcancel", O), o.value === V.Enter && $.type !== "animationcancel" && F.classList.remove(...y));
-    }, b = () => {
-      const $ = C();
-      $.classList.add(...y), $.addEventListener("animationend", O), $.addEventListener("animationcancel", O);
-    };
-    e.pauseOnFocusLoss && v(), b();
-  }
-  function S() {
-    if (!C())
-      return;
-    const y = () => {
-      const b = C();
-      b.removeEventListener("animationend", y), s.collapse ? at(b, u, s.collapseDuration) : u();
-    }, O = () => {
-      const b = C();
-      o.value = V.Exit, b && (b.className += " ".concat(d.value), b.addEventListener("animationend", y));
-    };
-    n.value || (a.value ? y() : setTimeout(O));
-  }
-  function C() {
-    return e.toastRef.value;
-  }
-  function v() {
-    document.hasFocus() || h(), window.addEventListener("focus", _), window.addEventListener("blur", h);
-  }
-  function I() {
-    window.removeEventListener("focus", _), window.removeEventListener("blur", h);
-  }
-  function _() {
-    (!e.loading.value || e.isLoading === void 0) && (t.value = !0);
-  }
-  function h() {
-    t.value = !1;
-  }
-  function R(y) {
-    y && (y.stopPropagation(), y.preventDefault()), n.value = !1;
-  }
-  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.watchEffect)(S), (0,vue__WEBPACK_IMPORTED_MODULE_0__.watchEffect)(() => {
-    const y = k();
-    n.value = y.findIndex((O) => O.toastId === s.toastId) > -1;
-  }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.watchEffect)(() => {
-    e.isLoading !== void 0 && (e.loading.value ? h() : _());
-  }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.onMounted)(m), (0,vue__WEBPACK_IMPORTED_MODULE_0__.onUnmounted)(() => {
-    e.pauseOnFocusLoss && I();
-  }), {
-    isIn: n,
-    isRunning: t,
-    hideToast: R,
-    eventHandlers: p
-  };
-}
-const st = /* @__PURE__ */ (0,vue__WEBPACK_IMPORTED_MODULE_0__.defineComponent)({
-  name: "ToastItem",
-  inheritAttrs: !1,
-  props: Ee,
-  // @ts-ignore
-  setup(e) {
-    const t = (0,vue__WEBPACK_IMPORTED_MODULE_0__.ref)(), n = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => !!e.isLoading), a = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => e.progress !== void 0 && e.progress !== null), o = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => tt(e)), s = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => ["".concat(r.CSS_NAMESPACE, "__toast"), "".concat(r.CSS_NAMESPACE, "__toast-theme--").concat(e.theme), "".concat(r.CSS_NAMESPACE, "__toast--").concat(e.type), e.rtl ? "".concat(r.CSS_NAMESPACE, "__toast--rtl") : void 0, e.toastClassName || ""].filter(Boolean).join(" ")), {
-      isRunning: u,
-      isIn: E,
-      hideToast: d,
-      eventHandlers: p
-    } = ot({
-      toastRef: t,
-      loading: n,
-      done: () => {
-        A.remove(e.toastId);
-      },
-      ...me(e.transition),
-      ...e
-    });
-    return () => (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("div", (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)({
-      id: e.toastId,
-      class: s.value,
-      style: e.toastStyle || {},
-      ref: t,
-      "data-testid": "toast-item-".concat(e.toastId),
-      onClick: (m) => {
-        e.closeOnClick && d(), e.onClick && e.onClick(m);
-      }
-    }, p.value), [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("div", {
-      role: e.role,
-      "data-testid": "toast-body",
-      class: "".concat(r.CSS_NAMESPACE, "__toast-body ").concat(e.bodyClassName || "")
-    }, [o.value != null && (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("div", {
-      "data-testid": "toast-icon-".concat(e.type),
-      class: ["".concat(r.CSS_NAMESPACE, "__toast-icon"), e.isLoading ? "" : "".concat(r.CSS_NAMESPACE, "--animate-icon ").concat(r.CSS_NAMESPACE, "__zoom-enter")].join(" ")
-    }, [G(o.value) ? (0,vue__WEBPACK_IMPORTED_MODULE_0__.h)((0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(o.value), {
-      theme: e.theme,
-      type: e.type
-    }) : B(o.value) ? o.value({
-      theme: e.theme,
-      type: e.type
-    }) : o.value]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)("div", {
-      "data-testid": "toast-content"
-    }, [G(e.content) ? (0,vue__WEBPACK_IMPORTED_MODULE_0__.h)((0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(e.content), {
-      toastProps: (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(e),
-      closeToast: d,
-      data: e.data
-    }) : B(e.content) ? e.content({
-      toastProps: (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(e),
-      closeToast: d,
-      data: e.data
-    }) : e.dangerouslyHTMLString ? (0,vue__WEBPACK_IMPORTED_MODULE_0__.h)("div", {
-      innerHTML: e.content
-    }) : e.content])]), (e.closeButton === void 0 || e.closeButton === !0) && (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(We, {
-      theme: e.theme,
-      closeToast: (m) => {
-        m.stopPropagation(), m.preventDefault(), d();
-      }
-    }, null), G(e.closeButton) ? (0,vue__WEBPACK_IMPORTED_MODULE_0__.h)((0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(e.closeButton), {
-      closeToast: d,
-      type: e.type,
-      theme: e.theme
-    }) : B(e.closeButton) ? e.closeButton({
-      closeToast: d,
-      type: e.type,
-      theme: e.theme
-    }) : null, (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(Qe, {
-      className: e.progressClassName,
-      style: e.progressStyle,
-      rtl: e.rtl,
-      theme: e.theme,
-      isIn: E.value,
-      type: e.type,
-      hide: e.hideProgressBar,
-      isRunning: u.value,
-      autoClose: e.autoClose,
-      controlledProgress: a.value,
-      progress: e.progress,
-      closeToast: e.isLoading ? void 0 : d
-    }, null)]);
-  }
-});
-let U = 0;
-function Se() {
-  typeof window > "u" || (U && window.cancelAnimationFrame(U), U = window.requestAnimationFrame(Se), Q.lastUrl !== window.location.href && (Q.lastUrl = window.location.href, A.clear()));
-}
-const rt = /* @__PURE__ */ (0,vue__WEBPACK_IMPORTED_MODULE_0__.defineComponent)({
-  name: "ToastifyContainer",
-  inheritAttrs: !1,
-  props: Ee,
-  // @ts-ignore
-  setup(e) {
-    const t = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => e.containerId), n = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => c[t.value] || []), a = (0,vue__WEBPACK_IMPORTED_MODULE_0__.computed)(() => n.value.filter((o) => o.position === e.position));
-    return (0,vue__WEBPACK_IMPORTED_MODULE_0__.onMounted)(() => {
-      typeof window < "u" && e.clearOnUrlChange && window.requestAnimationFrame(Se);
-    }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.onUnmounted)(() => {
-      typeof window < "u" && U && (window.cancelAnimationFrame(U), Q.lastUrl = "");
-    }), () => (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, [a.value.map((o) => {
-      const {
-        toastId: s = ""
-      } = o;
-      return (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(st, (0,vue__WEBPACK_IMPORTED_MODULE_0__.mergeProps)({
-        key: s
-      }, o), null);
-    })]);
-  }
-});
-let X = !1;
-function ve() {
-  const e = [];
-  return k().forEach((n) => {
-    const a = document.getElementById(n.containerId);
-    a && !a.classList.contains(K) && e.push(n);
-  }), e;
-}
-function lt(e) {
-  const t = ve().length, n = e != null ? e : 0;
-  return n > 0 && t + N.items.length >= n;
-}
-function it(e) {
-  lt(e.limit) && !e.updateId && N.items.push({
-    toastId: e.toastId,
-    containerId: e.containerId,
-    toastContent: e.content,
-    toastProps: e
-  });
-}
-function L(e, t, n = {}) {
-  if (X)
-    return;
-  n = Y(je(), {
-    type: t
-  }, (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(n)), (!n.toastId || typeof n.toastId != "string" && typeof n.toastId != "number") && (n.toastId = ye()), n = {
-    ...n,
-    content: e,
-    containerId: n.containerId || String(n.position)
-  };
-  const a = Number(n == null ? void 0 : n.progress);
-  return a < 0 && (n.progress = 0), a > 1 && (n.progress = 1), n.theme === "auto" && (n.theme = Ge()), it(n), Q.lastUrl = window.location.href, n.multiple ? N.items.length ? n.updateId && j(e, n) : j(e, n) : (X = !0, l.clearAll(void 0, !1), setTimeout(() => {
-    j(e, n);
-  }, 0), setTimeout(() => {
-    X = !1;
-  }, 390)), n.toastId;
-}
-const l = (e, t) => L(e, g.DEFAULT, t);
-l.info = (e, t) => L(e, g.DEFAULT, {
-  ...t,
-  type: g.INFO
-});
-l.error = (e, t) => L(e, g.DEFAULT, {
-  ...t,
-  type: g.ERROR
-});
-l.warning = (e, t) => L(e, g.DEFAULT, {
-  ...t,
-  type: g.WARNING
-});
-l.warn = l.warning;
-l.success = (e, t) => L(e, g.DEFAULT, {
-  ...t,
-  type: g.SUCCESS
-});
-l.loading = (e, t) => L(e, g.DEFAULT, Y(t, {
-  isLoading: !0,
-  autoClose: !1,
-  closeOnClick: !1,
-  closeButton: !1,
-  draggable: !1
-}));
-l.dark = (e, t) => L(e, g.DEFAULT, Y(t, {
-  theme: M.DARK
-}));
-l.remove = (e) => {
-  e ? A.dismiss(e) : A.clear();
-};
-l.clearAll = (e, t) => {
-  A.clear(e, t);
-};
-l.isActive = (e) => {
-  let t = !1;
-  return t = ve().findIndex((a) => a.toastId === e) > -1, t;
-};
-l.update = (e, t = {}) => {
-  setTimeout(() => {
-    const n = ke(e);
-    if (n) {
-      const a = (0,vue__WEBPACK_IMPORTED_MODULE_0__.toRaw)(n), {
-        content: o
-      } = a, s = {
-        ...a,
-        ...t,
-        toastId: t.toastId || e,
-        updateId: ye()
-      }, u = s.render || o;
-      delete s.render, L(u, s.type, s);
-    }
-  }, 0);
-};
-l.done = (e) => {
-  l.update(e, {
-    isLoading: !1,
-    progress: 1
-  });
-};
-l.promise = ut;
-function ut(e, {
-  pending: t,
-  error: n,
-  success: a
-}, o) {
-  var m, S, C;
-  let s;
-  const u = {
-    ...o || {},
-    autoClose: !1
-  };
-  t && (s = ae(t) ? l.loading(t, u) : l.loading(t.render, {
-    ...u,
-    ...t
-  }));
-  const E = {
-    autoClose: (m = o == null ? void 0 : o.autoClose) != null ? m : !0,
-    closeOnClick: (S = o == null ? void 0 : o.closeOnClick) != null ? S : !0,
-    closeButton: (C = o == null ? void 0 : o.autoClose) != null ? C : null,
-    isLoading: void 0,
-    draggable: null,
-    delay: 100
-  }, d = (v, I, _) => {
-    if (I == null) {
-      l.remove(s);
-      return;
-    }
-    const h = {
-      type: v,
-      ...E,
-      ...o,
-      data: _
-    }, R = ae(I) ? {
-      render: I
-    } : I;
-    return s ? l.update(s, {
-      ...h,
-      ...R,
-      isLoading: !1
-    }) : l(R.render, {
-      ...h,
-      ...R,
-      isLoading: !1
-    }), _;
-  }, p = B(e) ? e() : e;
-  return p.then((v) => {
-    d("success", a, v);
-  }).catch((v) => {
-    d("error", n, v);
-  }), p;
-}
-l.POSITION = D;
-l.THEME = M;
-l.TYPE = g;
-l.TRANSITIONS = Ne;
-const dt = {
-  install(e, t = {}) {
-    ct(t);
-  }
-};
-typeof window < "u" && (window.Vue3Toastify = dt);
-function ct(e = {}) {
-  const t = Y(fe, e);
-  ze(t);
-}
-
 
 
 /***/ })
